+++ biaws-cli/src/cli/admin/instance/update.js
@@ -1,0 +1,1 @@
+export { default } from "../../../commands/instance/update.js";